+++ src/configs/vendor-rules/style.ts
@@ -3,7 +3,7 @@
 // License: MIT (see LICENSE-eslint-config-airbnb.md file)
 // Permalink: https://github.com/airbnb/javascript/blob/c25bce83be4db06e6a221d79686c485cd2ed5d5d/packages/eslint-config-airbnb-base/rules/style.js
 
-import { Linter } from 'eslint';
+import type { Linter } from 'eslint';
 
 const rules = {
   // require function expressions to have a name