+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 [//]: # (https://keepachangelog.com/en/1.1.0/)
 
+## [0.0.4] - 2025-12-27
+
+- remove `arrow-body-style` from 'recommeded': not relevant for typed code
+- add `@typescript-eslint/consistent-type-imports` to `recommeded`: it helps with three-shaking
+
 ## [0.0.3] - 2025-12-25
 
 - disable recommended `mobx/missing-make-observable`: this rule useless with modern decorators syntax